refactor : 기존에 API 결과 Binding 전략을 appConfig에서 전략 구현체를 일일이 지정 해주는 방식 -> ApiQueryBinder에 map 형태로 전략 구현체를 가지고 전략 패턴을 적용하는 방식으로 변경.
