+++ src/main/java/com/scaling/libraryservice/search/service/strategy/EngKorNaturalSt.java
@@ -2,7 +2,6 @@ package com.scaling.libraryservice.search.service.strategy;
 
 import com.scaling.libraryservice.search.entity.Book;
 import com.scaling.libraryservice.search.repository.BookRepository;
-import com.scaling.libraryservice.search.service.SelectStrategy;
 import com.scaling.libraryservice.search.util.TitleQuery;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
