--- src/test/java/com/scaling/libraryservice/commons/api/service/KakaoBookProviderTest.java
@@ -1,7 +1,0 @@
-package com.scaling.libraryservice.commons.api.service;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class KakaoBookProviderTest {
-
-}
